backend/requirements/common.txt
# django-cors-headers
pillow
requests
bleach
asgiref==3.6.0
bleach==6.0.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-taggit==3.1.0
djangorestframework==3.14.0
idna==3.4
Pillow==9.4.0
Pygments==2.14.0
pytz==2022.7.1
requests==2.28.2
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
webencodings==0.5.1
django-cors-headers==3.14.0
boto3
django-storages